deps: update build scripts

+++ src/services/user_services/option_parser.cpp
@@ -14,8 +14,8 @@ ForwardOptions::ForwardOptions() : from(), to() {}
 ForwardOptions OptionParser::ParseForwardOptions(
     const std::string& option, boost::system::error_code& ec) {
   ForwardOptions forward_options;
-  const char *p;
-  char *end;
+  const char* p;
+  char* end;
   unsigned long val;
   std::string::size_type index, e;
 
@@ -32,7 +32,7 @@ ForwardOptions OptionParser::ParseForwardOptions(
     ec.assign(::error::invalid_argument, ::error::get_ssf_category());
     return {};
   }
-  forward_options.to.port = val;
+  forward_options.to.port = static_cast<uint16_t>(val);
 
   e = index;
   index = option.rfind(':', e - 1);
@@ -59,7 +59,7 @@ ForwardOptions OptionParser::ParseForwardOptions(
     ec.assign(::error::invalid_argument, ::error::get_ssf_category());
     return {};
   }
-  forward_options.from.port = val;
+  forward_options.from.port = static_cast<uint16_t>(val);
 
   return forward_options;
 }
@@ -68,8 +68,8 @@ Endpoint OptionParser::ParseListeningOption(const std::string& option,
                                             boost::system::error_code& ec) {
   Endpoint listening_option;
   unsigned long val;
-  const char *p;
-  char *end = NULL;
+  const char* p;
+  char* end = NULL;
   std::string::size_type index = option.find(':');
 
   if (index != std::string::npos) {
@@ -85,7 +85,7 @@ Endpoint OptionParser::ParseListeningOption(const std::string& option,
     ec.assign(::error::invalid_argument, ::error::get_ssf_category());
     return {};
   }
-  listening_option.port = val;
+  listening_option.port = static_cast<uint16_t>(val);
 
   return listening_option;
 }
@@ -93,7 +93,7 @@ Endpoint OptionParser::ParseListeningOption(const std::string& option,
 uint16_t OptionParser::ParsePort(const std::string& port,
                                  boost::system::error_code& ec) {
   unsigned long val;
-  char *end = NULL;
+  char* end = NULL;
 
   val = std::strtoul(port.c_str(), &end, 10);
   if (*end != '\0' || val > 65535) {
@@ -101,9 +101,8 @@ uint16_t OptionParser::ParsePort(const std::string& port,
     return 0;
   }
 
-  return val;
+  return static_cast<uint16_t>(val);
 }
-
 
 }  // services
 }  // ssf
